+++ src/gui/keybinds_reference.rs
@@ -10,11 +10,11 @@ const SCALED_KEY_PADDING: f32 = 0.0;
 const MIN_KEY_PADDING: f32 = 4.0;
 
 pub fn build(ui: &mut egui::Ui, app: &mut App) {
-    let prefs = app.prefs.gui.keybinds_reference;
+    let prefs = app.prefs.info.keybinds_reference;
 
     ui.scope(|ui| {
         let bg_fill = &mut ui.visuals_mut().widgets.noninteractive.bg_fill;
-        let alpha = app.prefs.gui.keybinds_reference.opacity;
+        let alpha = app.prefs.info.keybinds_reference.opacity;
         *bg_fill = bg_fill.linear_multiply(alpha);
 
         let mut areas = vec![MAIN_KEYS];
@@ -70,13 +70,13 @@ pub fn build(ui: &mut egui::Ui, app: &mut App) {
     });
 
     ui.collapsing("Settings", |ui| {
-        let default_prefs = crate::preferences::DEFAULT_PREFS.gui.keybinds_reference;
+        let default_prefs = crate::preferences::DEFAULT_PREFS.info.keybinds_reference;
 
         let mut changed = false;
 
         let r = ui.add(super::util::WidgetWithReset {
             label: "Opacity",
-            value: &mut app.prefs.gui.keybinds_reference.opacity,
+            value: &mut app.prefs.info.keybinds_reference.opacity,
             reset_value: default_prefs.opacity,
             reset_value_str: format!("{:.0}%", default_prefs.opacity * 100.0,),
             make_widget: super::util::make_percent_drag_value,
@@ -84,18 +84,18 @@ pub fn build(ui: &mut egui::Ui, app: &mut App) {
         changed |= r.changed();
 
         let r = ui.checkbox(
-            &mut app.prefs.gui.keybinds_reference.function,
+            &mut app.prefs.info.keybinds_reference.function,
             "Function keys",
         );
         changed |= r.changed();
 
         let r = ui.checkbox(
-            &mut app.prefs.gui.keybinds_reference.navigation,
+            &mut app.prefs.info.keybinds_reference.navigation,
             "Navigation keys",
         );
         changed |= r.changed();
 
-        let r = ui.checkbox(&mut app.prefs.gui.keybinds_reference.numpad, "Numpad");
+        let r = ui.checkbox(&mut app.prefs.info.keybinds_reference.numpad, "Numpad");
         changed |= r.changed();
 
         app.prefs.needs_save |= changed;