+++ src/preferences/info.rs
@@ -6,4 +6,15 @@ use crate::puzzle::TwistMetric;
 #[serde(default)]
 pub struct InfoPreferences {
     pub metric: TwistMetric,
+    pub keybinds_reference: KeybindsReferencePreferences,
+}
+
+#[derive(Serialize, Deserialize, Debug, Default, Copy, Clone)]
+#[serde(default)]
+pub struct KeybindsReferencePreferences {
+    pub function: bool,
+    pub navigation: bool,
+    pub numpad: bool,
+
+    pub opacity: f32,
 }