+++ src/render/structs.rs
@@ -1,7 +1,9 @@
 //! Structs shared between the CPU and GPU (vertices, uniforms, etc.).
 
-use cgmath::{Point3, Vector3};
+use cgmath::{Point2, Point3, Vector3};
 use smallvec::SmallVec;
+
+use super::util::IterCyclicPairsExt;
 
 #[repr(C)]
 #[derive(Debug, Default, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
@@ -22,6 +24,8 @@ impl RgbaVertex {
 
 #[derive(Debug)]
 pub(super) struct ProjectedStickerGeometry {
+    pub sticker_id: usize,
+
     pub verts: Box<[Point3<f32>]>,
     pub min_bound: Point3<f32>,
     pub max_bound: Point3<f32>,
@@ -31,6 +35,14 @@ pub(super) struct ProjectedStickerGeometry {
 
     pub outlines: Box<[[u16; 2]]>,
     pub outline_color: [f32; 4],
+    pub outline_size: f32,
+}
+impl ProjectedStickerGeometry {
+    pub(super) fn contains_point(&self, point: Point2<f32>) -> bool {
+        self.front_polygons
+            .iter()
+            .any(|polygon| polygon.contains_point(point))
+    }
 }
 
 #[derive(Debug, Clone)]
@@ -41,6 +53,21 @@ pub(super) struct Polygon {
     pub normal: Vector3<f32>,
 
     pub color: [f32; 4],
+}
+impl Polygon {
+    fn contains_point(&self, point: Point2<f32>) -> bool {
+        // println!("{:?}", (self.min_bound.x, self.min_bound.y));
+        self.min_bound.x <= point.x
+            && self.min_bound.y <= point.y
+            && point.x <= self.max_bound.x
+            && point.y <= self.max_bound.y
+            && self
+                .verts
+                .iter()
+                .map(|v| cgmath::point2(v.x, v.y))
+                .cyclic_pairs()
+                .all(|(a, b)| (b - a).perp_dot(point - a) <= 0.0)
+    }
 }
 
 #[repr(C)]